Lay groundwork for `ToggleMarked`
`ToggleMarked` is a new command that allows windows to stay rendering, even if they're hidden off in another screen. This allows programs like OBS and Discord to capture the window, even when its out of view.

At the moment, the window gets marked when it's toggled, but the window still appears on other tags.